Fix teams stupidity by making two swallows
Teams sucks.  When it starts, it throws up a load window that has the
same class as the main window.  So the load screen was getting
swallowed, and the main window ended up on whatever workspace was
active.

Creating two identical windows allows the first one to swallow the load
screen, which then disappears when it dies, and the second one swallows
the main window, which is the desired result.